requirements.txt
aiosql==8.0
beautifulsoup4==4.12.3
Flask==2.3.2
Flask_Login==0.6.2
Flask_WTF==1.1.1
Jinja2==3.1.2
psycopg==3.1.9
psycopg-binary==3.1.9
pytest==7.4.4
python-dotenv==1.0.0
pytz==2023.3.post1
Requests==2.31.0
scheduler==0.8.5
selenium==4.16.0
waitress==2.1.2
Werkzeug==2.3.6
WTForms==3.0.1
yoyo_migrations==8.2.0
